[add]: BaseLabeledImageDataset.
