+++ clib/training/dataset/__init__.py
@@ -1,7 +1,9 @@
 from clib.training.dataset import yolo
 from clib.training.dataset import unifieddataset
 from clib.training.dataset import xmldataset
+from clib.training.dataset import basedataset
 
 from clib.training.dataset.yolo import YoloPreprocessedDataset
 from clib.training.dataset.unifieddataset import UnifiedLabeledImageDataset
 from clib.training.dataset.xmldataset import XMLLabeledImageDataset
+from clib.training.dataset.basedataset import BaseLabeledImageDataset
